Clean up _copy_type_metadata (#11156)
This PR moves index-related logic from `Frame._copy_type_metadata` into the `IndexedFrame` override. It also removes all reference to the index parameter from the corresponding methods of various index types.

Authors:
  - Vyas Ramasubramani (https://github.com/vyasr)

Approvers:
  - Ashwin Srinath (https://github.com/shwina)
  - Charles Blackmon-Luca (https://github.com/charlesbluca)

URL: https://github.com/rapidsai/cudf/pull/11156
