+++ python/cudf/cudf/core/_base_index.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 import pickle
 import warnings
 from functools import cached_property
-from typing import Any, Set
+from typing import Any, Set, TypeVar
 
 import pandas as pd
 
@@ -64,6 +64,8 @@ Int64Index([1, 2, 3], dtype='int64')
 Float64Index([1.0, 2.0, 3.0], dtype='float64')
 """
 
+BaseIndexT = TypeVar("BaseIndexT", bound="BaseIndex")
+
 
 class BaseIndex(Serializable):
     """Base class for all cudf Index types."""
@@ -99,6 +101,9 @@ class BaseIndex(Serializable):
 
     def __contains__(self, item):
         return item in self._values
+
+    def _copy_type_metadata(self: BaseIndexT, other: BaseIndexT) -> BaseIndexT:
+        raise NotImplementedError
 
     def get_level_values(self, level):
         """
